Move ElectionError.tsx
To the router errors folder
