+++ src/router/index.tsx
@@ -2,13 +2,13 @@ import { ExtendedSDKClient } from '@vocdoni/extended-sdk'
 import { useClient } from '@vocdoni/react-providers'
 import { lazy } from 'react'
 import { createBrowserRouter, RouteObject, RouterProvider } from 'react-router-dom'
-import { BlockNotFound } from '~src/router/errors/BlockNotFound'
 import { RoutePath } from '~constants'
 import Layout from '~src/layout/Default'
+import { BlockNotFound } from '~src/router/errors/BlockNotFound'
+import { ElectionError } from '~src/router/errors/ElectionError'
 import Error404 from '~src/router/errors/Error404'
 import RouteError from '~src/router/errors/RouteError'
 import { SuspenseLoader } from '~src/router/SuspenseLoader'
-import { ElectionError } from '~src/router/ElectionError'
 
 const Home = lazy(() => import('~pages/Home'))
 const Block = lazy(() => import('~pages/block'))
